+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     - Bumped grpc-java to 1.76.0 to address CVE-2025-58056 from transient dep
     - Added `workloadIdentityImpersonationPath` config option for `authenticator=WORKLOAD_IDENTITY` allowing workloads to authenticate as a different identity through transitive service account impersonation (snowflakedb/snowflake-jdbc#2348)
     - Added support for authentication as a different identity through transitive IAM role impersonation for AWS (snowflakedb/snowflake-jdbc#2364)
+    - Add AWS identity detection with ARN validation (snowflakedb/snowflake-jdbc#2379)
 
 - v3.27.0
     - Added the `changelog.yml` GitHub workflow to ensure changelog is updated on release PRs.